Move `execute_task()` to a dedicated module (#3701)
# Description

Multiple executors use the `execute_task()` function, so moving it to
its own module improves code organization and reusability.

Also removed MPI-related code from `execute_task()`, as it's specific to
the HTEX.

## Type of change

- Code maintenance/cleanup
