+++ parsl/executors/high_throughput/process_worker_pool.py
@@ -23,6 +23,7 @@ import zmq
 
 from parsl import curvezmq
 from parsl.app.errors import RemoteExceptionWrapper
+from parsl.executors.execute_task import execute_task
 from parsl.executors.high_throughput.errors import WorkerLost
 from parsl.executors.high_throughput.mpi_prefix_composer import (
     VALID_LAUNCHERS,
@@ -35,7 +36,7 @@ from parsl.executors.high_throughput.mpi_resource_management import (
 from parsl.executors.high_throughput.probe import probe_addresses
 from parsl.multiprocessing import SpawnContext
 from parsl.process_loggers import wrap_with_logs
-from parsl.serialize import serialize, unpack_res_spec_apply_message
+from parsl.serialize import serialize
 from parsl.version import VERSION as PARSL_VERSION
 
 HEARTBEAT_CODE = (2 ** 32) - 1
@@ -597,40 +598,6 @@ def _init_mpi_env(mpi_launcher: str, resource_spec: Dict):
     nodes_for_task = node_list.split(',')
     logger.info(f"Launching task on provisioned nodes: {nodes_for_task}")
     update_resource_spec_env_vars(mpi_launcher=mpi_launcher, resource_spec=resource_spec, node_info=nodes_for_task)
-
-
-def execute_task(bufs: bytes):
-    """Deserialize the buffer and execute the task.
-    Returns the result or throws exception.
-    """
-    f, args, kwargs, resource_spec = unpack_res_spec_apply_message(bufs, copy=False)
-
-    for varname in resource_spec:
-        envname = "PARSL_" + str(varname).upper()
-        os.environ[envname] = str(resource_spec[varname])
-
-    # We might need to look into callability of the function from itself
-    # since we change it's name in the new namespace
-    prefix = "parsl_"
-    fname = prefix + "f"
-    argname = prefix + "args"
-    kwargname = prefix + "kwargs"
-    resultname = prefix + "result"
-
-    code = "{0} = {1}(*{2}, **{3})".format(resultname, fname,
-                                           argname, kwargname)
-
-    user_ns = locals()
-    user_ns.update({
-        '__builtins__': __builtins__,
-        fname: f,
-        argname: args,
-        kwargname: kwargs,
-        resultname: resultname
-    })
-
-    exec(code, user_ns, user_ns)
-    return user_ns.get(resultname)
 
 
 @wrap_with_logs(target="worker_log")
